+++ .github/ci/perl-lib/LoxBerry/System.pm
@@ -8,6 +8,9 @@ our $lbpbindir = '/opt/loxberry/bin/plugins/smartmeter-v2';
 our $lbpconfigdir = '/opt/loxberry/config/plugins/smartmeter-v2';
 our $lbptemplatedir = '/opt/loxberry/templates/plugins/smartmeter-v2';
 our $lbplogdir = '/opt/loxberry/log/plugins/smartmeter-v2';
+our $lbsconfigdir = '/opt/loxberry/config/system';
+our $lbslogdir = '/opt/loxberry/log/system';
+our $lbstmpfslogdir = '/opt/loxberry/log/system_tmpfs';
 
 sub import {
 	my $caller = caller;
@@ -18,6 +21,9 @@ sub import {
 	*{"${caller}::lbpconfigdir"} = \$lbpconfigdir;
 	*{"${caller}::lbptemplatedir"} = \$lbptemplatedir;
 	*{"${caller}::lbplogdir"} = \$lbplogdir;
+	*{"${caller}::lbsconfigdir"} = \$lbsconfigdir;
+	*{"${caller}::lbslogdir"} = \$lbslogdir;
+	*{"${caller}::lbstmpfslogdir"} = \$lbstmpfslogdir;
 }
 
 sub pluginversion { return '0.0.0'; }